generalized hf zip download fn
